functions.php: Override parent theme's font URL function
Twenty Sixteen uses fonts that don't have good coverage for Cyrillic
script. We override them and use Fira Sans and Fira Mono instead.

This includes a simple, verbatim port of all the parent's CSS rules
that are incompatible. For example, we obviously override all references
to the old font-families, but also, Fira Sans doesn't have a 900
weight, so when replacing Merriweather we need to use 700 weight.

Signed-off-by: Alan Orth <alan.orth@gmail.com>